+++ keyboards/handwired/tractyl_manuform_ct/5x6_right/5x6_right.c
@@ -14,48 +14,71 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "5x6_right.h"
-
 /*
-led_config_t g_led_config = { 
-// Key Matrix to LED Index
-{
-    { 0, 1, 2, 3, 4, 5 },
-    { 6, 7, 8, 9, 10, 11 },
-    { 12, 13, 14, 15, 16, 17 },
-    { 18, 19, 20, 21, 22 },
-    { 23, 24, 25, 26 }, 
-    { 27, 28, 29, 30, 31, 32 }, 
-    { 33, 34, 35, 36, 37, 38 }, 
-    { 39, 40, 41, 42, 43, 44 }, 
-    { 45, 46, 47, 48, 49, 50 }, 
-    { 51, 52, 53, 54, 55, 56 }
-}, 
-// LED Index to Physical Position
-{
-    { 102,   0 }, {  81,   0 }, {  61,   0 }, {  41,   0 }, {  20,   0 }, {   0,   0 },
-    { 102,  16 }, {  81,  16 }, {  61,  16 }, {  41,  16 }, {  20,  16 }, {   0,  16 },
-    { 102,  32 }, {  81,  32 }, {  61,  32 }, {  41,  32 }, {  20,  32 }, {   0,  32 },
-    { 102,  48 }, {  81,  48 }, {  61,  48 }, {  41,  48 }, {  20,  48 }, {   0,  48 },
-    { 102,  64 }, {  81,  64 }, {  61,  64 },
-    { 224,   0 }, { 204,   0 }, { 183,   0 }, { 163,   0 }, { 143,   0 }, { 122,   0 },
-    { 224,  16 }, { 204,  16 }, { 183,  16 }, { 163,  16 }, { 143,  16 }, { 122,  16 },
-    { 224,  32 }, { 204,  32 }, { 183,  32 }, { 163,  32 }, { 143,  32 }, { 122,  32 },
-    { 224,  48 }, { 204,  48 }, { 183,  48 }, { 163,  48 }, { 143,  48 }, { 122,  48 },
-    { 224,  64 }, { 204,  64 }, { 183,  64 }, { 163,  64 }, { 143,  64 }, { 122,  64 }
-},
-// LED type bit mask - 4 is per-key 
-{
-        4, 4, 4, 4, 4, 4,
-        4, 4, 4, 4, 4, 4,
-        4, 4, 4, 4, 4, 4,
-                4, 4, 4, 4, 4,
-                4, 4, 4, 4,
-            4, 4, 4, 4, 4, 4,
-            4, 4, 4, 4, 4, 4,
-            4, 4, 4, 4, 4, 4,
-      4, 4, 4, 4, 4, 4,
-      4, 4, 4, 4, 4, 4
-} 
-};
+    This file contains the logical key matrix's specific C code.
 */
+
+#include "5x6_right.h"
+#include "christrotter.h"
+// clang-format off
+    /* 
+     * ,-----------------------------------------,                                           ,-----------------------------------------,
+     * |  1,1 |  1,2 |  1,3 |  1,4 |  1,5 |  1,6 |                                           |  1,1 |  1,2 |  1,3 |  1,4 |  1,5 |  1,6 |
+     * |------+------+------+------+------+------|                                           |-------------+------+------+------+------|
+     * |  2,1 |  2,2 |  2,3 |  2,4 |  2,5 |  2,6 |                               (TRACKBALL) |  2,1 |  2,2 |  2,3 |  2,4 |  2,5 |  2,6 |
+     * |------+------+------+------+------|------|------|------|------,        |------|------|------|------+------+------+------+------|
+     * |  3,1 |  3,2 |  3,3 |  3,4 |  3,5 |  3,6 |  4,3 |  4,4 |  4,5 |        |  4,1 |  4,2 |  3,1 |  3,2 |  3,3 |  3,4 |  3,5 |  3,6 |
+     * |------+------+------+------,------,------,------,------,------,      ,------,------,------,------,-----------------------------,
+     *                             |      |      |      |      |             |      |      |      |      |        ,------,
+     *                             |  4,1 |  4,2 |  5,3 |  5,4 |             |  4,3 |  4,4 |  4,5 |  4,6 |        |  5,4 |
+     *                             |      |      |      |      |             |      |      |      |      | ,------+------+-----,
+     *                             '------+------+------+------'             '------'------+------+------' |  5,3 |------| 5,6 |
+     *                                    '------+------'                                  '------+------' '------+------+-----'
+     *                                    |  5,1 |  5,2 |                                  |  5,1 |  5,2 |        |  5,5 |
+     *                                    '------+------'                                  '------+------'        '------'
+     */
+// clang-format on
+#if defined(RGB_MATRIX_ENABLE)
+    led_config_t g_led_config = { 
+    // Key Matrix to LED Index
+    {
+        { 0, 1, 2, 3, 4, 5 },
+        { 6, 7, 8, 9, 10, 11 },
+        { 12, 13, 14, 15, 16, 17 },
+        { 18, 19, 20, 21, 22 },
+        { 23, 24, 25, 26 }, 
+        { 27, 28, 29, 30, 31, 32 }, 
+        { 33, 34, 35, 36, 37, 38 }, 
+        { 39, 40, 41, 42, 43, 44 }, 
+        { 45, 46, 47, 48, 49, 50 }, 
+        { 51, 52, 53, 54, 55, 56 }
+    }, 
+    // LED Index to Physical Position
+    {
+        { 102,   0 }, {  81,   0 }, {  61,   0 }, {  41,   0 }, {  20,   0 }, {   0,   0 },
+        { 102,  16 }, {  81,  16 }, {  61,  16 }, {  41,  16 }, {  20,  16 }, {   0,  16 },
+        { 102,  32 }, {  81,  32 }, {  61,  32 }, {  41,  32 }, {  20,  32 }, {   0,  32 },
+        { 102,  48 }, {  81,  48 }, {  61,  48 }, {  41,  48 }, {  20,  48 }, {   0,  48 },
+        { 102,  64 }, {  81,  64 }, {  61,  64 },
+        { 224,   0 }, { 204,   0 }, { 183,   0 }, { 163,   0 }, { 143,   0 }, { 122,   0 },
+        { 224,  16 }, { 204,  16 }, { 183,  16 }, { 163,  16 }, { 143,  16 }, { 122,  16 },
+        { 224,  32 }, { 204,  32 }, { 183,  32 }, { 163,  32 }, { 143,  32 }, { 122,  32 },
+        { 224,  48 }, { 204,  48 }, { 183,  48 }, { 163,  48 }, { 143,  48 }, { 122,  48 },
+        { 224,  64 }, { 204,  64 }, { 183,  64 }, { 163,  64 }, { 143,  64 }, { 122,  64 }
+    },
+    // LED type bit mask - 4 is per-key 
+    {
+            4, 4, 4, 4, 4, 4,
+            4, 4, 4, 4, 4, 4,
+            4, 4, 4, 4, 4, 4,
+                    4, 4, 4, 4, 4,
+                    4, 4, 4, 4,
+                4, 4, 4, 4, 4, 4,
+                4, 4, 4, 4, 4, 4,
+                4, 4, 4, 4, 4, 4,
+        4, 4, 4, 4, 4, 4,
+        4, 4, 4, 4, 4, 4
+    } 
+    };
+#endif
+
