--- users/christrotter/keymap.c
@@ -1,84 +1,0 @@
-#include QMK_KEYBOARD_H
-
-#define _BASE 0
-#define _SYMBOLS 0
-
-enum custom_keycodes {
-  BASE = SAFE_RANGE,
-  SYMBOLS
-};
-
-/*
-#define _RAISE 1
-#define _LOWER 2
-
-#define SFT_ESC  SFT_T(KC_ESC)
-#define CTL_BSPC CTL_T(KC_BSPC)
-#define ALT_SPC  ALT_T(KC_SPC)
-#define SFT_ENT  SFT_T(KC_ENT)
-
-#define KC_ML KC_MS_LEFT
-#define KC_MR KC_MS_RIGHT
-#define KC_MU KC_MS_UP
-#define KC_MD KC_MS_DOWN
-#define KC_MB1 KC_MS_BTN1
-#define KC_MB2 KC_MS_BTN2
-
-#define RAISE MO(_RAISE)
-#define LOWER MO(_LOWER)
-*/
-    /*  18 + 18 + 9L + (8+4)R = 57 keys
-     * ,-----------------------------------------,                                           ,-----------------------------------------,
-     * |  1,1 |  1,2 |  1,3 |  1,4 |  1,5 |  1,6 |                                           |  1,1 |  1,2 |  1,3 |  1,4 |  1,5 |  1,6 |
-     * |------+------+------+------+------+------|                                           |-------------+------+------+------+------|
-     * |  2,1 |  2,2 |  2,3 |  2,4 |  2,5 |  2,6 |                               (TRACKBALL) |  2,1 |  2,2 |  2,3 |  2,4 |  2,5 |  2,6 |
-     * |------+------+------+------+------|------|------|------|------,        |------|------|------|------+------+------+------+------|
-     * |  3,1 |  3,2 |  3,3 |  3,4 |  3,5 |  3,6 |  4,3 |  4,4 |  4,5 |        |  4,1 |  4,2 |  3,1 |  3,2 |  3,3 |  3,4 |  3,5 |  3,6 |
-     * |------+------+------+------,------,------,------,------,------,      ,------,------,------,------,-----------------------------,
-     *                             |      |      |      |      |             |      |      |      |      |        ,------,
-     *                             |  4,1 |  4,2 |  5,3 |  5,4 |             |  4,3 |  4,4 |  4,5 |  4,6 |        |  5,4 |
-     *                             |      |      |      |      |             |      |      |      |      | ,------+------+-----,
-     *                             '------+------+------+------'             '------'------+------+------' |  5,3 |------| 5,6 |
-     *                                    '------+------'                                  '------+------' '------+------+-----'
-     *                                    |  5,1 |  5,2 |                                  |  5,1 |  5,2 |        |  5,5 |
-     *                                    '------+------'                                  '------+------'        '------'
-     */
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-    /* Base (qwerty)
-     * ,-----------------------------------------,                                           ,-----------------------------------------,
-     * |  TAB |   q  |   w  |   e  |   r  |   t  |                                           |   y  |   u  |   i  |   o  |   p  |  -   |
-     * |------+------+------+------+------+------|                                           |-------------+------+------+-------------|
-     * |      |   a  |   s  |   d  |   f  |   g  |                               (TRACKBALL) |   h  |   j  |   k  |   l  |   '  |  ;   |
-     * |------+------+------+------+------|------|------|------|------,        |------|------|------|------+------+------+-------------|
-     * |      |   z  |   x  |   c  |   v  |   b  | ESC  | BACK | FWD  |        | PLAY | CTRL |   n  |   m  |   ,  |   .  |   /  | PLAY |
-     * |------+------+------+------,------,------,------,------,------,      ,------,------,------,------,------+------+------+--------,
-     *                             |  BS  |  DEL | CMD  |  LT  |             | CMD  |  LT  |ENTER |      |        ,------,
-     *                             |      |      |      |      |             |      |      |  +   |SPACE |        |  up  |
-     *                             |      |      |      |      |             |      |      |CMDent|      | ,------+------+-----,
-     *                             '------+------+------+------'             '------'------+------+------' | left |------|right|
-     *                                    '------+------'                                  '------+------' '------+------+-----'
-     *                                    | SHIFT| ALT  |                                  |  <   |  >   |        | down |
-     *                                    '------+------'                                  '------+------'        '------'
-     */
-[_BASE] = LAYOUT( \
-    KC_TAB,    KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,                               KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_MINUS,
-    KC_LSFT,   KC_A,    KC_S,    KC_D,    KC_F,    KC_G,                               KC_H,    KC_J,    KC_K,    KC_L,    KC_QUOT, KC_SCLN,
-    KC_ESC,    KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,                               KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLASH, KC_P,
-                KC_BSPC, KC_DEL,  KC_ESC, KC_LBRC, KC_RBRC,        KC_MPLY, KC_LCTL, KC_LGUI, KC_R, KC_ENTER, KC_SPACE,
-                                KC_LSFT, KC_LALT, KC_LGUI, KC_R,        KC_COMM, KC_DOT, KC_LEFT, KC_UP, KC_DOWN, KC_RIGHT
-),
-[_SYMBOLS] = LAYOUT( \
-    KC_TAB,    KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,                               KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_MINUS,
-    KC_LSFT,   KC_A,    KC_S,    KC_D,    KC_F,    KC_G,                               KC_H,    KC_J,    KC_K,    KC_L,    KC_QUOT, KC_SCLN,
-    KC_ESC,    KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,                               KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLASH, KC_P,
-                KC_BSPC, KC_DEL,  KC_ESC, KC_LBRC, KC_RBRC,        KC_MPLY, KC_LCTL, KC_LGUI, KC_R, KC_ENTER, KC_SPACE,
-                                KC_LSFT, KC_LALT, KC_LGUI, KC_R,        KC_COMM, KC_DOT, KC_LEFT, KC_UP, KC_DOWN, KC_RIGHT
-)
-
-};
-
-void persistent_default_layer_set(uint16_t default_layer) {
-    eeconfig_update_default_layer(default_layer);
-    default_layer_set(default_layer);
-}
